+++ common/math.cpp
@@ -19,7 +19,7 @@ const double DEGREES_TO_RADIANS = 1.745329251994e-02;
 // Returns a vector containing { closest point x, closest point y, u }
 // Note: You can check whether the closest point was on the initial line segment by
 // checking whether u is between 0 and 1.
-vector<double> closest_point_on_line(int x1, int y1, int x2, int y2, int x3, int y3) {
+vector<double> closest_point_on_line(double x1, double y1, double x2, double y2, double x3, double y3) {
 	double distp2p1 = Point::distance(Point(x1, y1), Point(x2, y2));
 	if (distp2p1 == 0) {
 		cerr << "Error: Distance between start and end of line is 0.";
