+++ common/math.hpp
@@ -30,7 +30,7 @@ static inline uint32_t toPow2(uint32_t num) {
 	return num;
 }
 
-std::vector<double> closest_point_on_line(int x1, int y1, int x2, int y2, int x3, int y3);
+std::vector<double> closest_point_on_line(double x1, double y1, double x2, double y2, double x3, double y3);
 
 // Get a value in the range [0,360)
 double get_normalized_angle(double angle);
